add purchase order

+++ spring-rest-api/src/main/java/com/rapidtech/springrestapi/repository/CustomerRepo.java
@@ -4,5 +4,5 @@ import com.rapidtech.springrestapi.entity.CustomerEntity;
 import com.rapidtech.springrestapi.model.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepo extends JpaRepository<CustomerEntity, Integer> {
+public interface CustomerRepo extends JpaRepository<CustomerEntity, Long> {
 }
